+++ server/api/recipe.delete.ts
@@ -1,8 +1,9 @@
 import { Recipe } from '../dao/models';
 
 export default defineEventHandler(async (event) => {
+  const userId = event.context.userId;
   const query = getQuery(event);
   const id = query.id;
-  return await Recipe.destroy({where: {id: Number(id)}})
+  return await Recipe.destroy({where: {id: Number(id), userId}})
 })
